requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
Authlib==0.15.3
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==2.0.1
